// src/components/ui-kit/components/confetti/types.ts
export interface ConfettiProps {
  isActive?: boolean;
  duration?: number;
  interval?: number;
}

export interface StyleProps {
  fill?: string;
  backgroundColor?: any;
  width: number;
  height: number;
  borderRadius?: number;
  transform: string;
  left: string | number;
  top: string | number;
}